requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
pydantic>=2.10.1,<3.0.0
pydantic-settings==2.9.1
pydantic_core==2.33.1
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
sniffio==1.3.1
starlette==0.46.2
typing_extensions==4.13.2
urllib3==2.3.0
uvicorn==0.34.2
firebase-admin==6.5.0
google-cloud-firestore==2.21.0
slowapi==0.1.9
limits>=5.5.0
pytest==8.3.3
sentry-sdk==2.19.2
openpyxl==3.1.2
google-cloud-vision>=3.0.0
reportlab>=4.0.0
